Remove unneeded use of accelerator object in Tiles methods

+++ include/CLUEstering/DataFormats/alpaka/AssociationMap.hpp
@@ -27,7 +27,7 @@ namespace clue {
   template <typename TFunc>
   struct KernelComputeAssociations {
     template <typename TAcc>
-      requires std::is_invocable_r_v<uint32_t, TFunc, const TAcc&, size_t>
+      requires std::is_invocable_r_v<uint32_t, TFunc, const TAcc&, uint32_t>
     ALPAKA_FN_ACC void operator()(const TAcc& acc,
                                   size_t size,
                                   uint32_t* associations,
@@ -37,7 +37,7 @@ namespace clue {
       }
     }
     template <typename TAcc>
-      requires std::is_invocable_r_v<uint32_t, TFunc, size_t>
+      requires std::is_invocable_r_v<uint32_t, TFunc, uint32_t>
     ALPAKA_FN_ACC void operator()(const TAcc& acc,
                                   size_t size,
                                   uint32_t* associations,
